llm/requirements.txt
fastapi
langchain
uvicorn
openai
python-dotenv==1.0.0
pydantic==1.10.9